requirements.txt
altair==5.3.0
asteval==1.0.8
attrs==25.4.0
biopython==1.83
bleach==6.3.0
bokeh==3.4.3
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
dill==0.4.1
et_xmlfile==2.0.0
fonttools==4.61.1
fraggler==3.0.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
linkify-it-py==2.0.3
lmfit==1.3.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.15.0
numpy==1.26.4
openpyxl==3.1.5
packaging==26.0
pandas==2.2.2
pandas-flavor==0.6.0
panel==1.4.4
param==2.3.1
pillow==12.1.0
pip==26.0.1
plotly==6.5.2
pyparsing==3.3.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.5.0
scipy==1.13.1
setuptools==65.5.0
six==1.17.0
threadpoolctl==3.6.0
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
uncertainties==3.2.3
urllib3==2.6.3
webencodings==0.5.1
wheel==0.46.3
xarray<2025.7
xyzservices==2025.11.0
